Sync status via merge field trigger.
The endpoint that receives merge field updates can additionally update the status of the user.
Additional configuration has been added to be able to specify the field the status value is stored in, as well as whether to interpret the value as a boolean or string, for easy integration with "subscribed": true/false type set ups.